# Conflicts:
#	EZSwiftExtensionsTests/DictionaryTests.swift
#	EZSwiftExtensionsTests/EZSwiftExtensionsTestsNSURL.swift
#	EZSwiftExtensionsTests/EZSwiftExtensionsTestsString.swift
#	EZSwiftExtensionsTests/NSDateTests.swift
#	Sources/ArrayExtensions.swift
#	Sources/BlockButton.swift
#	Sources/BlockLongPress.swift
#	Sources/BlockPan.swift
#	Sources/BlockPinch.swift
#	Sources/BlockSwipe.swift
#	Sources/BlockTap.swift
#	Sources/BlockWebView.swift
#	Sources/BoolExtensions.swift
#	Sources/CGFloatExtensions.swift
#	Sources/DoubleExtensions.swift
#	Sources/EZSwiftFunctions.swift
#	Sources/ErrorTypeExtensions.swift
#	Sources/NSAttributedStringExtensions.swift
#	Sources/NSDateExtensions.swift
#	Sources/NSURLExtensions.swift
#	Sources/NSUserDefaultsExtension.swift
#	Sources/StringExtensions.swift
#	Sources/UIDeviceExtensions.swift
#	Sources/UIImageExtensions.swift
#	Sources/UISliderExtensions.swift
#	Sources/UIStoryboardExtensions.swift
#	Sources/UITextViewExtensions.swift
#	Sources/UIViewControllerExtensions.swift
#	Sources/UIViewExtensions.swift

+++ Sources/NSAttributedStringExtensions.swift
@@ -45,10 +45,9 @@ extension NSAttributedString {
     /// EZSE: Adds strikethrough attribute to NSAttributedString and returns it
     public func strikethrough() -> NSAttributedString {
         guard let copy = self.mutableCopy() as? NSMutableAttributedString else { return self }
-
         let range = (self.string as NSString).range(of: self.string)
         let attributes = [
-        NSStrikethroughStyleAttributeName: NSNumber(value: NSUnderlineStyle.styleSingle.rawValue as Int)]
+        NSStrikethroughStyleAttributeName: NSNumber(value: NSUnderlineStyle.styleSingle.rawValue)]
         copy.addAttributes(attributes, range: range)
 
         return copy
